Fixing comments, tests, locks, annotations
