+++ Tests/MongoSwiftTests/ClientSessionTests.swift
@@ -48,50 +48,4 @@ final class ClientSessionTests: MongoSwiftTestCase {
             expect(escapedSession?.active).to(beFalse())
         }
     }
-
-    func testCommandEventStreamClient() async throws {
-        try await self.withTestClient { client in
-            let commandStr: [String] = ["ping", "ping", "endSessions", "endSessions"]
-            let eventTypes: [EventType] = [
-                .commandStartedEvent,
-                .commandSucceededEvent,
-                .commandStartedEvent,
-                .commandSucceededEvent
-            ]
-            Task {
-                var i = 0
-                for try await event in client.commandEvents {
-                    expect(commandStr[i]).to(equal(event.commandName))
-                    expect(eventTypes[i]).to(equal(event.type))
-
-                    i += 1
-                }
-            }
-            try await client.db("admin").runCommand(["ping": 1])
-        }
-    }
-
-    func testSDAMEventStreamClient() async throws {
-        try await self.withTestClient { client in
-            Task {
-                var i = 0
-                var eventHandler: [EventType] = []
-                client.addSDAMEventHandler { event in
-                    if !event.isHeartbeatEvent {
-                        eventHandler.append(event.type)
-                    }
-                }
-
-                for try await event in client.sdamEvents {
-                    if !event.isHeartbeatEvent {
-                        expect(event.type).to(equal(eventHandler[i]))
-                        i += 1
-                    }
-                }
-            }
-
-            try await client.db("admin").runCommand(["ping": 1])
-            try await client.db("trialDB").collection("trialColl").insertOne(["hello": "world"])
-        }
-    }
 }
